Update Close component
