+++ components/atoms/close/components/index.js
@@ -1,0 +1,4 @@
+/**
+ * Close
+ */
+export * from './close'
